requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
actuator-msgs==0.0.1
ament-clang-format==0.12.12
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.1
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
annotated-types==0.7.0
anyio==4.9.0
attrdict==2.0.1
blinker==1.9.0
builtin-interfaces==1.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
composition-interfaces==1.2.1
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.51.0
controller-manager-msgs==2.51.0
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
dnspython==2.7.0
domain-coordinator==0.10.0
email_validator==2.2.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
fsspec==2025.7.0
generate-parameter-library-py==0.5.0
geometry-msgs==4.9.0
gps-msgs==2.1.0
graph-msgs==0.2.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
image-geometry==3.2.1
interactive-markers==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.8
laser-geometry==2.4.0
launch==1.0.9
launch-param-builder==0.1.1
launch-ros==0.19.9
launch-testing==1.0.9
launch-testing-ros==0.19.9
launch-xml==1.0.9
launch-yaml==1.0.9
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
message-filters==4.3.7
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
mpmath==1.3.0
nav-msgs==4.9.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
opencv-python==4.12.0.88
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
pyrealsense2==2.56.5.9235
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-qt-binding==1.1.2
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
requests==2.32.4
resource-retriever==3.1.3
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.20
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2controlcli==2.51.0
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2test==0.4.0
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.15.3
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
sentry-sdk==2.34.1
shape-msgs==4.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
srdfdom==2.0.7
sros2==0.10.6
starlette==0.47.2
statistics-msgs==1.2.1
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.13
tf2-kdl==0.25.13
tf2-msgs==0.25.13
tf2-py==0.25.13
tf2-ros-py==0.25.13
tf2-sensor-msgs==0.25.13
tf2-tools==0.25.13
theora-image-transport==2.5.3
topic-monitor==0.20.5
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
trajectory-msgs==4.9.0
triton==3.3.1
turtlesim==1.4.2
tuyapy==0.1.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.170
ultralytics-thop==2.0.14
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vision-msgs==4.1.1
visualization-msgs==4.9.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
xacro==2.0.13
